fix: proper error handling for the migration

+++ src/utils/exiftool.ts
@@ -1,11 +1,17 @@
 import { execFile } from "promisify-child-process";
-import fs from "fs/promises";
+import fs from "fs-extra";
 import tempfile from "tempfile";
+import { logger } from ".";
 
 export const updateImageDescription = async (
   file: string,
   description: string,
 ) => {
+  if (fs.existsSync(file) === false) {
+    logger.error(`File not found: ${file}`);
+    return { error: "File not found" };
+  }
+
   // save description to a temp file
   const argFile = tempfile(".argFile");
   await fs.writeFile(argFile, description);
@@ -19,4 +25,5 @@ export const updateImageDescription = async (
   await fs.unlink(argFile);
   await fs.copyFile(newFile, file);
   await fs.unlink(newFile);
+  return {};
 };
